refactor(API): wip: api route updates
- moved API constants to constants file
- moved dataset-route methods to datasets service
- added JSDocs
- changed method names in UI and workers
