+++ api/src/constants.js
@@ -118,6 +118,15 @@ const WORKFLOWS = {
   CANCEL_DATASET_UPLOAD: 'cancel_dataset_upload',
 };
 
+const DATASET_STATES = {
+  READY: 'READY',
+  ARCHIVED: 'ARCHIVED',
+  REGISTERED: 'REGISTERED',
+  FETCHED: 'FETCHED',
+  STAGED: 'STAGED',
+  DELETED: 'DELETED',
+};
+
 module.exports = {
   INCLUDE_FILES,
   INCLUDE_STATES,
@@ -128,4 +137,5 @@ module.exports = {
   DATASET_CREATE_METHODS,
   UPLOAD_STATUSES,
   WORKFLOWS,
+  DATASET_STATES,
 };
