+++ workers/workers/tasks/download.py
@@ -11,7 +11,7 @@ import workers.api as api
 import workers.config.celeryconfig as celeryconfig
 from workers.config import config
 from workers.exceptions import ValidationFailed
-from workers.dataset import get_bundle_staged_path, get_bundle_name
+from workers.dataset import get_bundle_staged_path, get_staged_bundle_name
 
 app = Celery("tasks")
 app.config_from_object(celeryconfig)
@@ -56,7 +56,7 @@ def setup_download(celery_task, dataset_id, **kwargs):
 
     download_dir = Path(config['paths']['download_dir']).resolve()
     download_path = download_dir / alias
-    bundle_download_path = download_dir / get_bundle_name(dataset) 
+    bundle_download_path = download_dir / get_staged_bundle_name(dataset)
 
     # remove if exists and create a symlink in download dir pointing to the staged path
     rm(download_path)
